Fix no-virtualization issues and reorder style to affect offsetTop
The style being mounted after the lineDiv caused all lineDivs to have zero height and thus the loop was infinite. I moved the styles up top and adjusted the indices in the update method so now lineDiv grows as expected with the style applied. Unfortunately it seems that its offsetTop won't update right after adding it to the shadow root, so I had to use requestAnimationFrame and make the render method async. I think there must be a better solution but I can't think of one now.